requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
distlib==0.3.8
Django==4.2.13
django-cors-headers==4.3.1
django-filter==24.2
django-taggit==5.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.3.0
filelock==3.14.0
idna==3.4
Markdown==3.6
patool==1.12
pillow==10.3.0
pipenv==2023.12.1
platformdirs==4.2.1
progress==1.6
psycopg==3.1.18
psycopg2==2.9.9
PyJWT==2.8.0
requests==2.30.0
rest-framework-simplejwt==0.0.2
setuptools==69.5.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.0.2
virtualenv==20.26.1
